make updated resource manager work successfully on the client side

+++ src/core/java/net/remmintan/mods/minefortress/core/dtos/professions/ProfessionCost.kt
@@ -3,13 +3,13 @@ package net.remmintan.mods.minefortress.core.dtos.professions
 import net.minecraft.item.Item
 import net.minecraft.network.PacketByteBuf
 
-data class ProfessionCost(val item: Item, val requiredAmount: Int, val totalAmount: Long) {
+data class ProfessionCost(val item: Item, val requiredAmount: Long, val totalAmount: Long) {
 
     val enoughItems: Boolean = totalAmount >= requiredAmount
 
     fun writeToPacketByteBuf(buf: PacketByteBuf) {
         buf.writeInt(Item.getRawId(item))
-        buf.writeInt(requiredAmount)
+        buf.writeLong(requiredAmount)
         buf.writeLong(totalAmount)
     }
 
@@ -17,7 +17,7 @@ data class ProfessionCost(val item: Item, val requiredAmount: Int, val totalAmou
 
         fun readFromPacketByteBuf(buf: PacketByteBuf): ProfessionCost {
             val item = Item.byRawId(buf.readInt())
-            val requiredAmount = buf.readInt()
+            val requiredAmount = buf.readLong()
             val totalAmount = buf.readLong()
 
             return ProfessionCost(item, requiredAmount, totalAmount)
